# openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: Go Gretzky is committed to offering you an outstanding range of world
    class hockey products
  title: GoGretzky API
  version: 1.0.0
servers:
- url: https://api.example.com
paths:
  /:
    get:
      operationId: root_get
      responses:
        "200":
          content:
            application/json:
              schema:
                example: Welcome to the GoGretzky API
                type: string
                x-content-type: application/json
          description: Welcome to the GoGretzky API
      summary: Welcome to the GoGretzky API
      x-accepts: application/json
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/address:
    get:
      operationId: user_address_get
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: Home address of the logged in user
      security:
      - bearerAuth: []
      summary: Get the home address of the logged in user
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-accepts: application/json
    post:
      operationId: user_address_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_address_body'
      responses:
        "200":
          description: Successfully added home address to the logged in user
      security:
      - bearerAuth: []
      summary: Add a home address to the logged in user
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-contentType: application/json
      x-accepts: application/json
  /user/child:
    post:
      operationId: user_child_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_child_body'
      responses:
        "200":
          description: Successfully added child to the logged in user
      security:
      - bearerAuth: []
      summary: Add a child to the logged in user
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-contentType: application/json
      x-accepts: application/json
  /user/child/{child_id}:
    get:
      operationId: user_child_child_id_get
      parameters:
      - description: Id of the child to get
        explode: false
        in: path
        name: child_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_child_body'
          description: Child information
      security:
      - bearerAuth: []
      summary: Get the child by id
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-accepts: application/json
  /user/email:
    get:
      operationId: user_email_get
      responses:
        "200":
          content:
            application/json:
              schema:
                example: user@example.com
                type: string
                x-content-type: application/json
          description: Email of the logged in user
      security:
      - bearerAuth: []
      summary: Get the email of the logged in user
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-accepts: application/json
  /user/invitation:
    post:
      operationId: invitation_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/invitation_create_body'
      responses:
        "200":
          description: Successfully sent out the invitation
      security:
      - bearerAuth: []
      summary: Create an invitation and sends out an email to the given email address
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-contentType: application/json
      x-accepts: application/json
  /user/invitation/{invitation_id}:
    get:
      operationId: invitation_get
      parameters:
      - description: Id of the invitation
        explode: false
        in: path
        name: invitation_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invitation_information_body'
          description: Invitation information
      security:
      - bearerAuth: []
      summary: Get the invitation information with the given ID
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-accepts: application/json
  /user/invitations:
    get:
      operationId: invitations_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/invitation_information_body'
                type: array
          description: Invitation information
      security:
      - bearerAuth: []
      summary: Get all the invitations of the user
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-accepts: application/json
  /user/subscription:
    get:
      operationId: user_subscription_get
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
          description: Subscription of the logged in user
      security:
      - bearerAuth: []
      summary: Get the subscription of the logged in user
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-accepts: application/json
    post:
      operationId: user_subscription_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_subscription_body'
      responses:
        "200":
          description: Successfully added subscription to the logged in user
      security:
      - bearerAuth: []
      summary: Add a subscription to the logged in user
      x-openapi-router-controller: openapi_server.controllers.default_controller
      x-contentType: application/json
      x-accepts: application/json
components:
  schemas:
    invitation_information_body:
      example:
        tenant_id: gid:abcdefghijklmno
        message_attributes:
        - gid:aakkkkaaakkkaa
        reference_id: gid:1111kkkkk1111kkkkk111
        accepted_by: gid:kkkkkkiiiiiikkkkkkk
        expire_time: 2000-01-23T04:56:07.000+00:00
        invite_at_time: 2000-01-23T04:56:07.000+00:00
        state: INVITATION_STATE_ACCEPTED
        invitee: xxx@xxx.xx
      properties:
        accepted_by:
          nullable: true
          type: string
        expire_time:
          format: date-time
          type: string
        invite_at_time:
          format: date-time
          type: string
        invitee:
          type: string
        message_attributes:
          items:
            $ref: '#/components/schemas/invitation_information_body_message_attributes_inner'
          nullable: true
          type: array
        reference_id:
          type: string
        state:
          enum:
          - INVITATION_STATE_INVALID
          - INVITATION_STATE_IN_FUTURE
          - INVITATION_STATE_PENDING
          - INVITATION_STATE_ACCEPTED
          - INVITATION_STATE_EXPIRED
          - INVITATION_STATE_CANCELLED
          type: string
        tenant_id:
          type: string
      title: invitation_information_body
      type: object
    invitation_create_body:
      example:
        invitee: xxx@xxx.xx
      properties:
        invitee:
          format: email
          type: string
      title: invitation_create_body
    user_address_body:
      properties:
        street:
          example: Main St
          title: street
          type: string
        number:
          example: 123
          title: number
          type: integer
        city:
          example: Anytown
          title: city
          type: string
        state:
          example: CA
          title: state
          type: string
        zip:
          example: "12345"
          title: zip
          type: string
        country:
          example: USA
          title: country
          type: string
      title: user_address_body
      type: object
    inline_response_200_1:
      example:
        from: 2022-01-01T00:00:00Z
        to: 2022-12-31T23:59:59Z
        sku:
        - SKU22005
        - SKU22005
        child: 123e4567-e89b-12d3-a456-426655440000
      properties:
        from:
          example: 2022-01-01T00:00:00Z
          format: date-time
          type: string
        to:
          example: 2022-12-31T23:59:59Z
          format: date-time
          type: string
        sku:
          items:
            example: SKU22005
            type: string
          type: array
        child:
          example: 123e4567-e89b-12d3-a456-426655440000
          format: uuid
          type: string
      title: inline_response_200_1
      type: object
    inline_response_200:
      example:
        zip: "12345"
        number: 123
        country: USA
        city: Anytown
        street: Main St
        state: CA
      properties:
        street:
          example: Main St
          type: string
        number:
          example: 123
          type: integer
        city:
          example: Anytown
          type: string
        state:
          example: CA
          type: string
        zip:
          example: "12345"
          type: string
        country:
          example: USA
          type: string
      title: inline_response_200
      type: object
    user_child_body:
      example:
        shoe_size: 36
        gender: male
        cloth_size: 122
        helmet_size: 52
        given_name: John
        year_of_birth: 2010
      properties:
        given_name:
          example: John
          type: string
        year_of_birth:
          example: 2010
          format: int32
          type: integer
        gender:
          example: male
          type: string
        cloth_size:
          example: 122
          format: int32
          type: integer
        shoe_size:
          example: 36
          format: int32
          type: integer
        helmet_size:
          example: 52
          format: int32
          type: integer
      title: user_child_body
      type: object
    user_subscription_body:
      properties:
        from:
          example: 2022-01-01T00:00:00Z
          format: date-time
          title: from
          type: string
        to:
          example: 2022-12-31T23:59:59Z
          format: date-time
          title: to
          type: string
        sku:
          items:
            example: SKU22005
            type: string
          title: sku
          type: array
        child:
          example: 123e4567-e89b-12d3-a456-426655440000
          format: uuid
          title: child
          type: string
      title: user_subscription_body
      type: object
    invitation_information_body_message_attributes_inner:
      oneOf:
      - type: string
      - type: integer
      - type: number
      title: invitation_information_body_message_attributes_inner
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller_.info_from_bearerAuth